Move basic attachment models into separate file
